use hook for closing banner logic
